+++ guillotina/api/ws.py
@@ -33,7 +33,7 @@ import ujson
     responses={
         "200": {
             "description": "The new token",
-            "schema": {"properties": {"token": {"type": "string", "required": True}}},
+            "content": {"application/json": {"schema": {"properties": {"token": {"type": "string"}}}}},
         }
     },
 )
@@ -46,7 +46,7 @@ import ujson
     responses={
         "200": {
             "description": "The new token",
-            "schema": {"properties": {"token": {"type": "string", "required": True}}},
+            "content": {"application/json": {"schema": {"properties": {"token": {"type": "string"}}}}},
         }
     },
 )