+++ guillotina/routes.py
@@ -15,12 +15,18 @@ class RoutePart:
         return self.name == other
 
     def __repr__(self):
-        return "<RoutePart {}>".format(self.name)
+        return f"<{self.__class__.__name__} {self.name}>"
+
+    def __str__(self):
+        return self.name
 
 
 class VariableRoutePart(RoutePart):
     def matches(self, other):
         return True
+
+    def __str__(self):
+        return "{" + self.name + "}"
 
 
 class Route: