+++ CHANGELOG.rst
@@ -7,14 +7,19 @@ CHANGELOG
 - Run _update_from_pytest_markers() after configuring db settings
   [masipcat]
 
+- Fix bug in traversal introduced when added support for asgi middlewares
+  [masipcat]
+
 - Fix value_deserializer() when field.key_type._type is None
   [masipcat]
 
 - Fix automatic type conversion on nested fields. Fixes #832
   [vangheem]
 
+
 6.0.0a6 (2020-01-13)
 --------------------
+
 - Fix bug on swagger with endpoints without explicit security declarations
   [jordic]